requirements.txt
Flask==3.1.0
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.36
